+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ... <!-- Add new lines here. -->
 - docs: Improve docs and add JSDoc comments to most methods and properties
 - feat: Mark the `media` helper as deprecated
+- feat: Add ` cssVal`` ` templater alias
 
 ## 0.5.13